pyproject.toml
[project]
name = "text-processor-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "python-nmap>=0.7.1",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "xmltodict>=1.0.4",
    "yagmail>=0.15.293",
]
